Pi3プログラムを作り直し、Pi3Sample→Pi3Appへ変更
